+++ hack/generate.sh
@@ -3,10 +3,11 @@
 set -eux
 
 bash $GOPATH/src/k8s.io/code-generator/generate-groups.sh deepcopy,client,informer,lister \
-    $PROJECT/pkg/hub/k8s/client $PROJECT/pkg/hub/k8s/apis \
+    $PROJECT/pkg/client $PROJECT/pkg/apis \
     kopilot:v1alpha1 \
     --go-header-file hack/boilerplate.go.txt
 
 go generate ./...
 
-controller-gen paths=./... crd rbac:roleName=kopilot-hub webhook
+controller-gen paths=./... crd webhook output:webhook:artifacts:config=config/kopilot-webhook
+controller-gen paths=./pkg/hub/... rbac:roleName=kopilot-hub output:rbac:artifacts:config=config/kopilot-hub